+++ examples/info.R
@@ -1,14 +1,18 @@
 source("../R/Api.R")
 library(jsonlite)
-library("optparse")
+library(optparse)
 
-option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
-              help="Rosette API key"), make_option(c("-u", "--url"), action="store", default=NA, type='character',
-              help="Rosette API url"))
-opt_parser = OptionParser(option_list=option_list)
-opt = parse_args(opt_parser)
+option_list <- list(
+  make_option(c("-k", "--key"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API key"),
+  make_option(c("-u", "--url"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API url"))
+opt_parser <- OptionParser(option_list = option_list)
+opt <- parse_args(opt_parser)
 
-if (is.na(opt$url)){
+if (is.na(opt$url)) {
    result <- api(opt$key, "info")
 } else {
    result <- api(opt$key, "info", FALSE, NULL, NULL, opt$url)