+++ examples/name_deduplication.R
@@ -1,26 +1,30 @@
 source("../R/Api.R")
 library(jsonlite)
-library("optparse")
+library(optparse)
 
-option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
-              help="Rosette API key"), make_option(c("-u", "--url"), action="store", default=NA, type='character',
-              help="Rosette API url"))
-opt_parser = OptionParser(option_list=option_list)
-opt = parse_args(opt_parser)
+option_list <- list(
+  make_option(c("-k", "--key"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API key"),
+  make_option(c("-u", "--url"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API url"))
+opt_parser <- OptionParser(option_list = option_list)
+opt <- parse_args(opt_parser)
 
 name_dedupe_data <- "Alice Terry,Alice Thierry,Betty Grable,Betty Gable,Norma Shearer,Norm Shearer,Brigitte Helm,Bridget Helem,Judy Holliday,Julie Halliday"
 
 
 text <- unlist(strsplit(name_dedupe_data, ","))
-targetLanguage <- rep("eng", length(text))
-targetScript <- rep("Latn", length(text))
+target_language <- rep("eng", length(text))
+target_script <- rep("Latn", length(text))
 
-names = data.frame(text, targetLanguage, targetScript)
+names <- data.frame(text, target_language, target_script)
 parameters <- list()
-parameters[[ "names" ]] <- names
-parameters[[ "threshold" ]] <- 0.75
+parameters[["names"]] <- names
+parameters[["threshold"]] <- 0.75
 
-if (is.na(opt$url)){
+if (is.na(opt$url)) {
    result <- api(opt$key, "name-deduplication", parameters)
 } else {
    result <- api(opt$key, "name-deduplication", parameters, NULL, NULL, opt$url)