+++ examples/name_similarity.R
@@ -1,21 +1,26 @@
 source("../R/Api.R")
 library(jsonlite)
-library("optparse")
+library(optparse)
 
-option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
-              help="Rosette API key"), make_option(c("-u", "--url"), action="store", default=NA, type='character',
-              help="Rosette API url"))
-opt_parser = OptionParser(option_list=option_list)
-opt = parse_args(opt_parser)
+option_list <- list(
+  make_option(c("-k", "--key"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API key"),
+  make_option(c("-u", "--url"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API url"))
+opt_parser <- OptionParser(option_list = option_list)
+opt <- parse_args(opt_parser)
+
 
 matched_name_data1 <- "Michael Jackson"
 matched_name_data2 <- "迈克尔·杰克逊"
 
 parameters <- list()
-parameters[[ "name1" ]] <- matched_name_data1
-parameters[[ "name2" ]] <- matched_name_data2
+parameters[["name1"]] <- matched_name_data1
+parameters[["name2"]] <- matched_name_data2
 
-if (is.na(opt$url)){
+if (is.na(opt$url)) {
    result <- api(opt$key, "name-similarity", parameters)
 } else {
    result <- api(opt$key, "name-similarity", parameters, NULL, NULL, opt$url)