+++ examples/relationships.R
@@ -1,19 +1,23 @@
 source("../R/Api.R")
 library(jsonlite)
-library("optparse")
+library(optparse)
 
-option_list = list( make_option(c("-k", "--key"), action="store", default=NA, type='character',
-              help="Rosette API key"), make_option(c("-u", "--url"), action="store", default=NA, type='character',
-              help="Rosette API url"))
-opt_parser = OptionParser(option_list=option_list)
-opt = parse_args(opt_parser)
+option_list <- list(
+  make_option(c("-k", "--key"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API key"),
+  make_option(c("-u", "--url"),
+    action = "store", default = NA, type = "character",
+    help = "Rosette API url"))
+opt_parser <- OptionParser(option_list = option_list)
+opt <- parse_args(opt_parser)
 
 relationships_text_data <- "FLIR Systems is headquartered in Oregon and produces thermal imaging, night vision, and infrared cameras and sensor systems.  According to the SEC’s order instituting a settled administrative proceeding, FLIR entered into a multi-million dollar contract to provide thermal binoculars to the Saudi government in November 2008.  Timms and Ramahi were the primary sales employees responsible for the contract, and also were involved in negotiations to sell FLIR’s security cameras to the same government officials.  At the time, Timms was the head of FLIR’s Middle East office in Dubai."
 
 parameters <- list()
-parameters[[ "content" ]] <- relationships_text_data
+parameters[["content"]] <- relationships_text_data
 
-if (is.na(opt$url)){
+if (is.na(opt$url)) {
    result <- api(opt$key, "relationships", parameters)
 } else {
    result <- api(opt$key, "relationships", parameters, NULL, NULL, opt$url)