+++ tests/testthat/test_api.R
@@ -1,6 +1,4 @@
-#R binding unit tests
-
-# mock all HTTP httr::GET requests
+# Mock all HTTP httr::GET requests
 test_that("httr::GET function mocks correctly", {
   with_mock(
     `httr::GET` = function(x, y) "mocked response",
@@ -10,7 +8,7 @@ test_that("httr::GET function mocks correctly", {
   )
 })
 
-# mock all HTTP httr::POST requests
+# Mock all HTTP httr::POST requests
 test_that("httr::POST functions mock correctly", {
   parameters <- jsonlite::toJSON(list())
   with_mock(
@@ -23,16 +21,25 @@ test_that("httr::POST functions mock correctly", {
 
 test_that("The function detects a multipart", {
   parameters <- list()
-  parameters[[ "content" ]] <- "Rosette Api"
-  parameters[[ "documentFile" ]] <- "true"
+  parameters[["content"]] <- "Rosette Api"
+  parameters[["documentFile"]] <- "true"
   expect_equal(check_for_multipart(parameters), TRUE)
 })
 
 test_that("The function creates a multipart", {
+  payload <- paste("--89dszpjalrbmlsor\r\n",
+                   "Content-Type: application/json\r\n",
+                   "Content-Disposition: mixed; name=\"request\"\r\n\r\n",
+                   "{}\r\n\r\n",
+                   "--89dszpjalrbmlsor\r\n",
+                   "Content-Type: text/plain\r\n",
+                   "Content-Disposition: mixed; name=\"content\";\r\n\r\n",
+                   "Rosette Api\r\n",
+                   "--89dszpjalrbmlsor--", sep = "")
   parameters <- list()
-  parameters[[ "content" ]] <- "Rosette Api"
-  parameters[[ "documentFile" ]] <- "true"
-  expect_equal(create_multipart(parameters), "--89dszpjalrbmlsor\r\nContent-Type: application/json\r\nContent-Disposition: mixed; name=\"request\"\r\n\r\n{}\r\n\r\n--89dszpjalrbmlsor\r\nContent-Type: text/plain\r\nContent-Disposition: mixed; name=\"content\";\r\n\r\nRosette Api\r\n--89dszpjalrbmlsor--")
+  parameters[["content"]] <- "Rosette Api"
+  parameters[["documentFile"]] <- "true"
+  expect_equal(create_multipart(parameters), payload)
 })
 
 test_that("Check error does not return an error", {
@@ -47,29 +54,55 @@ test_that("409 error is handled correctly", {
   expect_equal(error_check(value), value)
 })
 
+test_that("The parameters fulfill the address-similarity requirements", {
+  parameters <- list()
+  parameters[["address1"]] <- list("houseNumber" = "1600",
+                                   "road" = "Pennsylvania Ave",
+                                   "city" = "Washington")
+  parameters[["address2"]] <- list("houseNumber" = "1600",
+                                   "road" = "Pennsylvania Ave",
+                                   "city" = "Washington",
+                                   "state" = "D.C.",
+                                   "postcode" = "20500")
+  expect_equal(check_address_parameters(parameters), parameters)
+})
+
+test_that("The parameter check for address-similarity fails", {
+  parameters <- list()
+  parameters[["address1"]] <- list("houseNumber" = "1600",
+                                   "road" = "Pennsylvania Ave",
+                                   "city" = "Washington")
+  parameters[["notaddress2"]] <- list("houseNumber" = "1600",
+                                      "road" = "Pennsylvania Ave",
+                                      "city" = "Washington",
+                                      "state" = "D.C.",
+                                      "postcode" = "20500")
+  expect_error(check_address_parameters(parameters),
+               "Must specify both address1 and address2 parameters")
+})
+
 test_that("The parameters fulfill the name-translation requirements", {
   parameters <- list()
-  parameters[[ "name" ]] <- "fake name"
-  parameters[[ "targetLanguage" ]] <- "eng"
+  parameters[["name"]] <- "fake name"
+  parameters[["targetLanguage"]] <- "eng"
   expect_equal(check_names(parameters, "name-translation"), parameters)
 })
 
 test_that("The parameters fulfill the name-similarity requirements", {
   parameters <- list()
-  parameters[[ "name1" ]] <- "first name"
-  parameters[[ "name2" ]] <- "second name"
+  parameters[["name1"]] <- "first name"
+  parameters[["name2"]] <- "second name"
   expect_equal(check_names(parameters, "name-similarity"), parameters)
 })
 
 test_that("The parameters fulfill the name-deduplication requirements", {
   name <- c("John Smith", "Johnathon Smith", "Fred Jones")
-  target.language <- c("eng", "eng", "eng")
-  target.script <- c("Latn", "Latn", "Latn")
-
-  names <- data.frame(name, target.language, target.script)
+  target_language <- c("eng", "eng", "eng")
+  target_script <- c("Latn", "Latn", "Latn")
+  names <- data.frame(name, target_language, target_script)
   parameters <- list()
-  parameters[[ "names" ]] <- names
-  parameters[[ "threshold" ]] <- 0.75
+  parameters[["names"]] <- names
+  parameters[["threshold"]] <- 0.75
 
   expect_equal(check_deduplication(parameters, "name-deduplication"),
                parameters)
@@ -77,21 +110,21 @@ test_that("The parameters fulfill the name-deduplication requirements", {
 
 test_that("The parameters fulfill the transliteration requirements", {
   parameters <- list()
-  parameters[[ "content" ]] <- "Mongo only pawn in game of life"
+  parameters[["content"]] <- "Mongo only pawn in game of life"
 
   expect_equal(check_content_parameters(parameters), parameters)
 })
 
 test_that("A morphology endpoint is present", {
   parameters <- list()
-  parameters[[ "content" ]] <- "Rosette Api"
-  parameters[[ "morphology" ]] <- "complete"
+  parameters[["content"]] <- "Rosette Api"
+  parameters[["morphology"]] <- "complete"
   expect_equal(check_morphology(parameters), "morphology/complete")
 })
 
 test_that("The parameters object is valid", {
   parameters <- list()
-  parameters[[ "content" ]] <- "Rosette Api"
+  parameters[["content"]] <- "Rosette Api"
   expect_equal(check_content_parameters(parameters), parameters)
 })